Gardening: Represent partitions as ints instead of shorts
Motivation
----------
Just because a partition identifier only uses 16 bits
in the wire protocol, it doesn't mean we have to use a special type to
represent it. This way, nobody has to wonder, "Why is this not an int?"

Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/138207
Tested-by: Build Bot <build@couchbase.com>

+++ examples/src/main/java/com/couchbase/client/dcp/examples/PrintIncomingChanges.java
@@ -59,7 +59,7 @@ public class PrintIncomingChanges {
           flowController.ack(event);
         }
         if (RollbackMessage.is(event)) {
-          final short partition = RollbackMessage.vbucket(event);
+          final int partition = RollbackMessage.vbucket(event);
           client.rollbackAndRestartStream(partition, RollbackMessage.seqno(event))
               .subscribe(new CompletableSubscriber() {
                 @Override
