+++ src/main/java/com/couchbase/client/dcp/message/DcpDeletionMessage.java
@@ -61,7 +61,7 @@ public enum DcpDeletionMessage {
     return MessageUtil.getCas(buffer);
   }
 
-  public static short partition(final ByteBuf buffer) {
+  public static int partition(final ByteBuf buffer) {
     return MessageUtil.getVbucket(buffer);
   }
