+++ src/main/java/com/couchbase/client/dcp/message/DcpMutationMessage.java
@@ -68,7 +68,7 @@ public enum DcpMutationMessage {
     return MessageUtil.getCas(buffer);
   }
 
-  public static short partition(final ByteBuf buffer) {
+  public static int partition(final ByteBuf buffer) {
     return MessageUtil.getVbucket(buffer);
   }
